requirements.txt
discord.py
pymongo
dnspython
#dotenv